renaming categories, adding storybook entry
